solutions/challenge2-safety/requirements.txt
paho-mqtt>=1.6.0
numpy>=1.26.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
